Review fixes: serde, GSS merge, json schema, position types, roadmap
- Add serde derives to all public glr-core types (cfg_attr)
- Fix GSS add_node merge: register new head on node reuse
- Re-export Production from glr-core crate root
- Add BLANK/PREC_DYNAMIC/Unknown variants to grammar JSON schema
- Fix glr-query circular module coupling (NodeRef in query.rs)
- Harmonize Token positions to u32 (was usize)
- Update TestLexer to use u32 positions
- Mark Phase 0 as completed in ROADMAP.md

+++ glr-lexer/src/token.rs
@@ -4,8 +4,8 @@ use glr_core::SymbolId;
 #[derive(Debug, Clone)]
 pub struct Token {
     pub kind: SymbolId,
-    pub start_byte: usize,
-    pub end_byte: usize,
+    pub start_byte: u32,
+    pub end_byte: u32,
 }
 
 /// Trait for lexers consumed by the GLR parser.
@@ -14,8 +14,8 @@ pub trait Lexer {
     fn next_token(&mut self, valid_symbols: &[bool]) -> Option<Token>;
 
     /// Current byte offset in the source.
-    fn cursor(&self) -> usize;
+    fn cursor(&self) -> u32;
 
     /// Reset to a specific byte offset.
-    fn reset_to(&mut self, byte_offset: usize);
+    fn reset_to(&mut self, byte_offset: u32);
 }
